+++ libs/cardano-ledger-core/CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## 1.22.0.0
 
-* Add `EbAnnouncement` (moved from `cardano-protocol`)
+* Rename `EraIndependentEb` to `EraIndependentEbReferences`
+* Add `EbReferencesAnnouncement` (moved and renamed from `EbAnnouncement` in `cardano-protocol`)
 * Add `headerBlockL`, `bodyBlockL`
 * Add `addVRFKeyHashOccurrence`, `removeVRFKeyHashOccurrence` and `populateVRFKeyHashes` to `Cardano.Ledger.State.CertState`
 * Remove `poolDistrDistrL`, `poolDistrTotalL`